factor helper methods to base class and update docstring
